Configure all units❤️ Implemented Recochet Property❤️ Minor refactoring❤️

+++ Assets/Scripts/GameCore/Battle/Data/GameProperty/ValuePercent.cs
@@ -3,6 +3,6 @@
 [Serializable]
 public struct ValuePercent
 {
-    public float value;
+    public decimal value;
     public float percent;
 }
